[6761] Extract the explorer subscription in a dedicated component
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6761
