+++ packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/explorer/ExplorerView.tsx
@@ -12,320 +12,41 @@
  *******************************************************************************/
 import {
   RepresentationLoadingIndicator,
-  Selection,
-  SelectionEntry,
-  useSelection,
   WorkbenchViewComponentProps,
   WorkbenchViewHandle,
 } from '@eclipse-sirius/sirius-components-core';
-import {
-  FilterBar,
-  GQLGetTreePathVariables,
-  GQLTree,
-  GQLTreeItem,
-  TreeFilter,
-  TreeToolBar,
-  TreeToolBarContext,
-  TreeToolBarContextValue,
-  TreeView,
-  useTreeFilters,
-  useTreePath,
-  useTreeSelection,
-} from '@eclipse-sirius/sirius-components-trees';
 import AccountTreeIcon from '@mui/icons-material/AccountTree';
 import Box from '@mui/material/Box';
-import { Theme } from '@mui/material/styles';
 import Typography from '@mui/material/Typography';
-import { ForwardedRef, forwardRef, useCallback, useContext, useEffect, useRef, useState } from 'react';
-import { makeStyles } from 'tss-react/mui';
-import { DuplicateObjectKeyboardShortcut } from './context-menu-contributions/duplicate-object/DuplicateObjectKeyboardShortcut';
+import { ForwardedRef, forwardRef, useEffect, useState } from 'react';
+import { ExplorerSubscriptionProvider } from './ExplorerSubscriptionProvider';
 import { ExplorerViewConfiguration, ExplorerViewState } from './ExplorerView.types';
-import { TreeDescriptionsMenu } from './TreeDescriptionsMenu';
 import { useExplorerDescriptions } from './useExplorerDescriptions';
-import { useExplorerSubscription } from './useExplorerSubscription';
-import { GQLTreeEventPayload, GQLTreeRefreshedEventPayload } from './useExplorerSubscription.types';
-import { useExplorerViewHandle } from './useExplorerViewHandle';
-
-const useStyles = makeStyles()((theme: Theme) => ({
-  treeView: {
-    display: 'grid',
-    gridTemplateColumns: 'auto',
-    gridTemplateRows: 'auto auto 1fr',
-    justifyItems: 'stretch',
-    overflow: 'auto',
-  },
-  treeFilter: {
-    paddingTop: theme.spacing(1),
-  },
-  treeContent: {
-    overflow: 'auto',
-  },
-}));
-
-const isTreeRefreshedEventPayload = (payload: GQLTreeEventPayload): payload is GQLTreeRefreshedEventPayload =>
-  payload && payload.__typename === 'TreeRefreshedEventPayload';
 
 export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewComponentProps>(
   (
     { editingContextId, id, initialConfiguration, readOnly }: WorkbenchViewComponentProps,
     ref: ForwardedRef<WorkbenchViewHandle>
   ) => {
-    const { classes: styles } = useStyles();
-
     const initialExplorerViewConfiguration: ExplorerViewConfiguration =
       initialConfiguration as unknown as ExplorerViewConfiguration;
+
     const [state, setState] = useState<ExplorerViewState>({
-      filterBar: false,
-      filterBarText: '',
-      filterBarTreeFiltering: false,
-      treeFilters: initialExplorerViewConfiguration?.activeTreeFilters ?? [],
-      activeTreeDescriptionId: initialExplorerViewConfiguration?.activeTreeDescriptionId ?? null,
-      expanded: {},
-      maxDepth: {},
-      tree: null,
-      selectedTreeItemIds: [],
-      singleTreeItemSelected: null,
-      selectionPivotTreeItemId: null,
+      initialActiveTreeDescriptionId: initialExplorerViewConfiguration?.activeTreeDescriptionId ?? null,
     });
 
-    // If we are requested to reveal the global selection, we need to compute the tree path to expand
-    const { getTreePath, data: treePathData } = useTreePath();
-
-    const applySelection = (selection: Selection) => {
-      const newSelectedTreeItemIds = selection.entries.map((entry) => entry.id);
-      setState((prevState) => ({
-        ...prevState,
-        selectedTreeItemIds: newSelectedTreeItemIds,
-      }));
-
-      if (state.tree && newSelectedTreeItemIds.length > 0) {
-        const variables: GQLGetTreePathVariables = {
-          editingContextId,
-          treeId: state.tree.id,
-          selectionEntryIds: newSelectedTreeItemIds,
-        };
-        getTreePath({ variables });
-      }
-    };
-
-    useExplorerViewHandle(id, state.tree?.id, state.treeFilters, state.activeTreeDescriptionId, applySelection, ref);
-
-    const treeToolBarContributionComponents = useContext<TreeToolBarContextValue>(TreeToolBarContext).map(
-      (contribution) => contribution.props.component
-    );
-    const activeTreeFilterIds = state.treeFilters.filter((filter) => filter.state).map((filter) => filter.id);
-
-    const { payload } = useExplorerSubscription(
-      editingContextId,
-      state.activeTreeDescriptionId,
-      activeTreeFilterIds,
-      state.expanded[state.activeTreeDescriptionId] ?? [],
-      state.maxDepth[state.activeTreeDescriptionId] ?? 1
-    );
-
-    useEffect(() => {
-      if (isTreeRefreshedEventPayload(payload)) {
-        setState((prevState) => ({ ...prevState, tree: payload.tree }));
-      }
-    }, [payload]);
-
-    const { explorerDescriptions } = useExplorerDescriptions(editingContextId);
+    const { loading, explorerDescriptions } = useExplorerDescriptions(editingContextId);
 
     useEffect(() => {
       if (explorerDescriptions && explorerDescriptions.length > 0) {
-        const expandedInitiated: { [key: string]: string[] } = {};
-        const maxDepthInitiated: { [key: string]: number } = {};
-        explorerDescriptions.forEach((explorerDescription) => {
-          expandedInitiated[explorerDescription.id] = [];
-          maxDepthInitiated[explorerDescription.id] = 1;
-        });
-
         setState((prevState) => ({
           ...prevState,
-          activeTreeDescriptionId: state.activeTreeDescriptionId ?? explorerDescriptions[0].id,
-          expanded: expandedInitiated,
-          maxDepth: maxDepthInitiated,
+          initialActiveTreeDescriptionId: state.initialActiveTreeDescriptionId ?? explorerDescriptions[0].id,
         }));
       }
     }, [explorerDescriptions]);
 
-    const { loading, treeFilters } = useTreeFilters(editingContextId, state.activeTreeDescriptionId || null);
-
-    useEffect(() => {
-      if (!loading) {
-        const allAvailableFilters: TreeFilter[] = treeFilters.map((gqlTreeFilter) => ({
-          id: gqlTreeFilter.id,
-          label: gqlTreeFilter.label,
-          state: gqlTreeFilter.defaultState,
-        }));
-        setState((prevState) => ({
-          ...prevState,
-          treeFilters: allAvailableFilters.map((availableFilter) => {
-            const existingFilter: TreeFilter = state.treeFilters.find((filter) => filter.id === availableFilter.id);
-            if (existingFilter) {
-              return {
-                ...availableFilter,
-                state: existingFilter.state,
-              };
-            } else {
-              return availableFilter;
-            }
-          }),
-        }));
-      }
-    }, [loading, treeFilters.map((treeFilter) => treeFilter.id).join()]);
-
-    const treeElement = useRef<HTMLDivElement>(null);
-    useEffect(() => {
-      const downHandler = (event) => {
-        if (
-          (event.ctrlKey === true || event.metaKey === true) &&
-          event.key === 'f' &&
-          event.target.tagName !== 'INPUT'
-        ) {
-          event.preventDefault();
-          setState((prevState) => {
-            return { ...prevState, filterBar: true, filterBarText: '', filterBarTreeFiltering: false };
-          });
-        }
-      };
-      const element = treeElement?.current;
-      if (element) {
-        element.addEventListener('keydown', downHandler);
-
-        return () => {
-          element.removeEventListener('keydown', downHandler);
-        };
-      }
-      return null;
-    }, [treeElement]);
-
-    const { selection, setSelection } = useSelection();
-    const { treeItemClick } = useTreeSelection();
-
-    const selectionKey: string = selection?.entries
-      .map((entry) => entry.id)
-      .sort()
-      .join(':');
-
-    const revealSelection = useCallback(() => {
-      if (state.tree && selection.entries.length > 0) {
-        const variables: GQLGetTreePathVariables = {
-          editingContextId,
-          treeId: state.tree.id,
-          selectionEntryIds: selection.entries.map((entry) => entry.id),
-        };
-        getTreePath({ variables });
-      }
-    }, [editingContextId, selectionKey, state.tree, getTreePath]);
-
-    useEffect(() => {
-      if (treePathData && treePathData.viewer?.editingContext?.treePath) {
-        setState((prevState) => {
-          const { expanded, maxDepth } = prevState;
-          const { treeItemIdsToExpand, maxDepth: expandedMaxDepth } = treePathData.viewer.editingContext.treePath;
-          const newExpanded: string[] = [...expanded[prevState.activeTreeDescriptionId]];
-
-          treeItemIdsToExpand?.forEach((itemToExpand) => {
-            if (!expanded[prevState.activeTreeDescriptionId].includes(itemToExpand)) {
-              newExpanded.push(itemToExpand);
-            }
-          });
-          return {
-            ...prevState,
-            selectedTreeItemIds: selection.entries.map((entry) => entry.id),
-            expanded: {
-              ...prevState.expanded,
-              [prevState.activeTreeDescriptionId]: newExpanded,
-            },
-            maxDepth: {
-              ...prevState.maxDepth,
-              [prevState.activeTreeDescriptionId]: Math.max(
-                expandedMaxDepth,
-                maxDepth[prevState.activeTreeDescriptionId]
-              ),
-            },
-          };
-        });
-      }
-    }, [treePathData]);
-
-    const onExpandedElementChange = (newExpandedIds: string[], newMaxDepth: number) => {
-      setState((prevState) => ({
-        ...prevState,
-        expanded: {
-          ...prevState.expanded,
-          [prevState.activeTreeDescriptionId]: newExpandedIds,
-        },
-        maxDepth: {
-          ...prevState.maxDepth,
-          [prevState.activeTreeDescriptionId]: Math.max(
-            newMaxDepth,
-            prevState.maxDepth[prevState.activeTreeDescriptionId]
-          ),
-        },
-      }));
-    };
-
-    let filterBar: JSX.Element = <div />;
-    if (state.filterBar) {
-      filterBar = (
-        <div className={styles.treeFilter}>
-          <FilterBar
-            onTextChange={(event) => {
-              const {
-                target: { value },
-              } = event;
-              setState((prevState) => {
-                return { ...prevState, filterBarText: value };
-              });
-            }}
-            onFilterButtonClick={(enabled) =>
-              setState((prevState) => ({
-                ...prevState,
-                filterBarTreeFiltering: enabled,
-              }))
-            }
-            onClose={() =>
-              setState((prevState) => {
-                return { ...prevState, filterBar: false, filterBarText: '', filterBarTreeFiltering: false };
-              })
-            }
-          />
-        </div>
-      );
-    }
-
-    const onTreeItemClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>, tree: GQLTree, item: GQLTreeItem) => {
-      var localSelection = treeItemClick(event, tree, item, state.selectedTreeItemIds, true);
-      setState((prevState) => ({
-        ...prevState,
-        selectedTreeItemIds: localSelection.selectedTreeItemIds,
-        singleTreeItemSelected: localSelection.singleTreeItemSelected,
-      }));
-      var globalSelection = treeItemClick(
-        event,
-        state.tree,
-        item,
-        selection.entries.map((entry) => entry.id),
-        true
-      );
-      setSelection({ entries: globalSelection.selectedTreeItemIds.map<SelectionEntry>((id) => ({ id })) });
-    };
-
-    const treeDescriptionSelector: JSX.Element = explorerDescriptions.length > 1 && (
-      <TreeDescriptionsMenu
-        treeDescriptions={explorerDescriptions}
-        activeTreeDescriptionId={state.activeTreeDescriptionId}
-        onTreeDescriptionChange={(treeDescription) =>
-          setState((prevState) => ({
-            ...prevState,
-            activeTreeDescriptionId: treeDescription.id,
-            tree: null,
-          }))
-        }
-      />
-    );
+    const initialTreeFilters = initialExplorerViewConfiguration?.activeTreeFilters ?? [];
 
     return (
       <Box sx={{ display: 'flex', flexDirection: 'column' }} data-testid="view-Explorer">
@@ -347,67 +68,19 @@ export const ExplorerView = forwardRef<WorkbenchViewHandle, WorkbenchViewCompone
             Explorer
           </Typography>
         </Box>
-        <Box className={styles.treeView} sx={{ flexGrow: 1, minHeight: 0 }} ref={treeElement}>
-          {!state.tree || loading ? (
-            <RepresentationLoadingIndicator />
-          ) : (
-            <>
-              <TreeToolBar
-                editingContextId={editingContextId}
-                readOnly={readOnly}
-                treeFilters={state.treeFilters}
-                onRevealSelection={revealSelection}
-                onTreeFilterMenuItemClick={(treeFilters) =>
-                  setState((prevState) => {
-                    return { ...prevState, treeFilters };
-                  })
-                }
-                onFilter={() => {
-                  setState((prevState) => {
-                    return !prevState.filterBar
-                      ? { ...prevState, filterBar: true, filterBarText: '', filterBarTreeFiltering: false }
-                      : { ...prevState, filterBar: false, filterBarText: '', filterBarTreeFiltering: false };
-                  });
-                }}
-                treeToolBarContributionComponents={treeToolBarContributionComponents}>
-                {treeDescriptionSelector}
-              </TreeToolBar>
-              <DuplicateObjectKeyboardShortcut
-                target={treeElement?.current}
-                editingContextId={editingContextId}
-                readOnly={readOnly}
-                selectedTreeItem={state.singleTreeItemSelected}
-                selectTreeItems={(selectedTreeItemIds: string[]) =>
-                  setState((prevState) => {
-                    return { ...prevState, selectedTreeItemIds };
-                  })
-                }>
-                {filterBar}
-                <div className={styles.treeContent}>
-                  <TreeView
-                    editingContextId={editingContextId}
-                    readOnly={readOnly}
-                    tree={state.tree}
-                    textToHighlight={state.filterBarText}
-                    textToFilter={state.filterBarTreeFiltering ? state.filterBarText : null}
-                    onExpandedElementChange={onExpandedElementChange}
-                    expanded={state.expanded[state.activeTreeDescriptionId]}
-                    maxDepth={state.maxDepth[state.activeTreeDescriptionId]}
-                    onTreeItemClick={onTreeItemClick}
-                    selectTreeItems={(selectedTreeItemIds: string[]) =>
-                      setState((prevState) => {
-                        return { ...prevState, selectedTreeItemIds };
-                      })
-                    }
-                    selectedTreeItemIds={state.selectedTreeItemIds}
-                    data-testid="explorer://"
-                    useTreePalette={state.tree.capabilities.useTreePalette}
-                  />
-                </div>
-              </DuplicateObjectKeyboardShortcut>
-            </>
-          )}
-        </Box>
+        {loading || !state.initialActiveTreeDescriptionId ? (
+          <RepresentationLoadingIndicator />
+        ) : (
+          <ExplorerSubscriptionProvider
+            id={id}
+            explorerRef={ref}
+            editingContextId={editingContextId}
+            readOnly={readOnly}
+            explorerDescriptions={explorerDescriptions}
+            initialActiveTreeDescriptionId={state.initialActiveTreeDescriptionId}
+            initialFilters={initialTreeFilters}
+          />
+        )}
       </Box>
     );
   }
