+++ packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/explorer/ExplorerView.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023, 2025 Obeo.
+ * Copyright (c) 2023, 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -11,20 +11,10 @@
  *     Obeo - initial API and implementation
  *******************************************************************************/
 import { WorkbenchViewConfiguration } from '@eclipse-sirius/sirius-components-core';
-import { GQLTree, GQLTreeItem, TreeFilter } from '@eclipse-sirius/sirius-components-trees';
+import { TreeFilter } from '@eclipse-sirius/sirius-components-trees';
 
 export interface ExplorerViewState {
-  filterBar: boolean;
-  filterBarText: string | null;
-  filterBarTreeFiltering: boolean;
-  treeFilters: TreeFilter[];
-  activeTreeDescriptionId: string | null;
-  expanded: { [key: string]: string[] };
-  maxDepth: { [key: string]: number };
-  tree: GQLTree | null;
-  selectedTreeItemIds: string[];
-  selectionPivotTreeItemId: string | null;
-  singleTreeItemSelected: GQLTreeItem | null;
+  initialActiveTreeDescriptionId: string | null;
 }
 
 export interface ExplorerViewConfiguration extends WorkbenchViewConfiguration {
